eliminate explicit (long) casts when making syscalls
this practice came from very early, before internal/syscall.h defined
macros that could accept pointer arguments directly and handle them
correctly. aside from being ugly and unnecessary, it looks like it
will be problematic when we add support for 32-bit ABIs on archs where
registers (and syscall arguments) are 64-bit, e.g. x32 and mips n32.

+++ src/time/timer_delete.c
@@ -10,5 +10,5 @@ int timer_delete(timer_t t)
 		__wake(&td->timer_id, 1, 1);
 		return 0;
 	}
-	return __syscall(SYS_timer_delete, (long)t);
+	return __syscall(SYS_timer_delete, t);
 }
